Replace while loops with for loops in Station and Signal Tabs

+++ src/openloco/windows/construction/OverheadTab.cpp
@@ -248,7 +248,7 @@ namespace openloco::ui::windows::construction::overhead
                 coord_t y = 0x2010;
 
                 auto rotCoord = rotate2DCoordinate({ x, y }, gCurrentRotation);
-                gfx::point_t screenPos = { rotCoord.y - rotCoord.x, ((rotCoord.x + rotCoord.y) >> 1) - 460 };
+                gfx::point_t screenPos = { static_cast<int16_t>(rotCoord.y - rotCoord.x), static_cast<int16_t>(((rotCoord.x + rotCoord.y) >> 1) - 460) };
 
                 screenPos.x -= (self->widgets[widx::image].width() / 2);
                 screenPos.y -= ((self->widgets[widx::image].width() / 2) + 16);
